Refactor: Change how the friendly KEY works.
I thought about it a bit. I've decided that the unordered_map API for resources is not a good idea for KEYs.
(maybe for ERFs also, will think about it.)

The common usage scenario of an KEY will be to access underlying resources from a BIF - you don't want to look
up the resource by name, generally. You want to get a big list of references resources. It makes sense to use a
vector instead in this case.